refactor(reviewer-bot): extract sweeper logic
